feat(primitives): added null object (#15)
### Context

According to the `ISO 32000-1:2008` specification, section `7.3.9`, the null object has a type and value that are unequal to those of any other object.

There shall be only one object of type null, denoted by the keyword `null`. An indirect object reference to a nonexistent object shall be treated the same as a null object. Specifying the null object as the value of a dictionary entry shall be equivalent to omitting the entry entirely.

### Acceptance criteria

1. Create a `PdfNull` struct.
2. Implement the `IPdfObject` interface.
3. Override the `GetHashCode` and `ToString` methods.
4. Create implicit operators.
6. Cover the primitive implementation with the unit and mutation tests.