feat(common-data-structures): added rectangle object (#152)
### Context

According to the `ISO 32000-1:2008` specification, section `7.9.5`, Rectangles are used to describe locations on a page and bounding boxes for a variety of objects. A rectangle shall be written as an array of four numbers giving the coordinates of a pair of diagonally opposite corners. 

Default form: $[ll_x\space ll_y\space ur_x\space ur_y]$, where the values correspond to:
1. lower-left x
2. lower-left y
3. upper-right x
4. upper-right y

![Rectangle form](https://user-images.githubusercontent.com/24962085/223048679-bda235d3-605a-4916-b4c8-e78a487c195b.png)

Even though the specification allows using any two diagonally opposite corners, this story will be focusing on supporting the above-described rectangle form.

### Acceptance criteria

1. This object must extend the #8 
2. The code must be covered with the unit tests